requirement.txt
flask
flask-script
flask-blueprint
flask-session
redis
flask-sqlalchemy
pymysql
baidu-aip
